backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bobo-backend"
version = "0.1.0"
description = "Bobo Agent backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "openai>=1.40",
  "qdrant-client>=1.10",
  "langgraph>=0.5,<0.6",
  "langgraph-checkpoint-postgres>=3,<4",
  "langchain-openai>=0.3",
  "langchain-mcp-adapters>=0.1",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7",
  "sqlmodel>=0.0.22",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.2",
  "psycopg-pool>=3.2",
  "python-multipart>=0.0.9",
  "cos-python-sdk-v5>=1.9.41",
  "fastmcp>=2.0",
  "httpx>=0.27",
]

[project.optional-dependencies]
all = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "alembic>=1.13",
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
